requirements.txt
sqlalchemy>=2.0.0
alembic>=1.0.0
python-dotenv>=0.19.0
typer>=0.9.0
rich>=13.0.0
